Add Dockerfile for building and running application
A new Dockerfile has been added to the project which sets up a multi-stage build process for building and running the application. The first stage uses a golang base image to build the application, while the second stage uses an alpine base image to run the built application. Port 6969 is exposed for communication with the application.